refactor: replace loader
