+++ httprunner/ext/make/__init__.py
@@ -7,7 +7,7 @@ from loguru import logger
 
 from httprunner import exceptions
 from httprunner.exceptions import TestCaseFormatError
-from httprunner.new_loader import load_testcase_file, load_folder_files
+from httprunner.loader import load_testcase_file, load_folder_files
 
 __TMPL__ = """# NOTICE: Generated By HttpRunner. DO'NOT EDIT!
 import unittest
@@ -96,9 +96,10 @@ def init_make_parser(subparsers):
     """ make testcases: parse command line options and run commands.
     """
     parser = subparsers.add_parser(
-        "make",
-        help="Convert YAML/JSON testcases to Python unittests.",
+        "make", help="Convert YAML/JSON testcases to Python unittests.",
     )
-    parser.add_argument("testcase_path", nargs="?", help="Specify YAML/JSON testcase path")
+    parser.add_argument(
+        "testcase_path", nargs="?", help="Specify YAML/JSON testcase path"
+    )
 
     return parser
