+++ httprunner/ext/make/make_test.py
@@ -3,13 +3,12 @@ from httprunner.ext.make import make_testcase, main_make
 
 
 class TestLoader(unittest.TestCase):
-
     def test_make_testcase(self):
         path = "examples/postman_echo/request_methods/request_with_variables.yml"
         testcase_python_path = make_testcase(path)
         self.assertEqual(
             testcase_python_path,
-            "examples/postman_echo/request_methods/request_with_variables_test.py"
+            "examples/postman_echo/request_methods/request_with_variables_test.py",
         )
 
     def test_make_testcase_folder(self):
@@ -17,5 +16,5 @@ class TestLoader(unittest.TestCase):
         testcase_python_list = main_make(path)
         self.assertIn(
             "examples/postman_echo/request_methods/request_with_functions_test.py",
-            testcase_python_list
+            testcase_python_list,
         )
