--- httprunner/loader/__init__.py
@@ -1,27 +1,0 @@
-"""
-HttpRunner loader
-
-- check: validate api/testcase/testsuite data structure with JSON schema
-- locate: locate debugtalk.py, make it's dir as project root path
-- load: load testcase files and relevant data, including debugtalk.py, .env, yaml/json api/testcases, csv, etc.
-- buildup: assemble loaded content to httprunner testcase/testsuite data structure
-
-"""
-
-from httprunner.loader.buildup import load_cases, load_project_data
-from httprunner.loader.check import is_test_path
-from httprunner.loader.load import load_csv_file, load_builtin_functions
-from httprunner.loader.locate import (
-    get_project_working_directory as get_pwd,
-    init_project_working_directory as init_pwd,
-)
-
-__all__ = [
-    "is_test_path",
-    "get_pwd",
-    "init_pwd",
-    "load_csv_file",
-    "load_builtin_functions",
-    "load_project_data",
-    "load_cases",
-]
