+++ httprunner/runner.py
@@ -8,7 +8,7 @@ from loguru import logger
 from httprunner import utils, exceptions
 from httprunner.client import HttpSession
 from httprunner.exceptions import ValidationFailure, ParamsError
-from httprunner.new_loader import load_project_meta, load_testcase_file
+from httprunner.loader import load_project_meta, load_testcase_file
 from httprunner.parser import build_url, parse_data, parse_variables_mapping
 from httprunner.response import ResponseObject
 from httprunner.schema import (
